gentle refactoring
