+++ digitalocean_plugin/tasks.py
@@ -1,5 +1,5 @@
 # #######
-# Copyright (c) 2014 GigaSpaces Technologies Ltd. All rights reserved
+# Copyright (c) 2015 GigaSpaces Technologies Ltd. All rights reserved
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,58 +12,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import os
 import digitalocean as ocean
 from cloudify import ctx
 from cloudify.decorators import operation
 from cloudify.exceptions import NonRecoverableError
 
-
-def load_token():
-    """ XXX
-    This will load a security token from a local file called token.txt
-    A token can be obtained from DigitalOcean by Registering a New Developer or Authorized Application.
-    :return: the security token, as a string
-    :raises: NonRecoverableError if token is not present
-    """
-    cwd = os.path.dirname(__file__)
-    token_path = os.path.join(cwd, 'token.txt')
-    if not os.path.isfile(token_path):
-        msg = 'Missing security token file "%s".' % token_path
-        ctx.logger.debug(msg)
-        raise NonRecoverableError(msg)
-    with open(token_path, 'r') as f:
-        return f.read()
-
-
-def available_images():
-    """ XXX
-    image specifiers are used to provision Droplets. Note: Not all images are available in all regions
-    :return: a list of available image specifiers
-    """
-    # TODO: Load from API: See https://developers.digitalocean.com/#list-all-images
-    return ['ubuntu-14-04-x64']
-
-
-def available_regions():
-    """ XXX
-    region specifiers are used to provision Droplets in a particular data center ('region').
-    Note: Not all images or options are available on all regions.
-    :return: a list of available region specifiers
-    """
-    # TODO Load from API: See https://developers.digitalocean.com/#list-all-regions
-    regions = ['nyc3', 'nyc1', 'nyc2']
-    return regions
-
-
-def available_slug_sizes(region):
-    """ XXX
-    :param region: region specifier for which to return slug sizes
-    :return: all available slug sizes
-    """
-    # TODO Load from API: See https://developers.digitalocean.com/#list-all-regions
-    sizes = ['512mb']
-    return sizes
+from security import load_token
+from utils import available_images, available_regions, available_slug_sizes, get_droplet, droplet_does_not_exist_for_operation
 
 
 def generate_droplet_name():
@@ -75,7 +30,7 @@ def generate_droplet_name():
 
 @operation
 def create(droplet_name=None, region=None, image=None, size_slug='512mb', backups=False, **_):
-    """ XXX
+    """
     Tell the API to create a droplet. Note that not all combinations of options are possible
     :param droplet_name: formal name
     :param region: region to choose
@@ -110,40 +65,6 @@ def create(droplet_name=None, region=None, image=None, size_slug='512mb', backup
     pass
 
 
-def get_droplet(droplet_id):
-    """ XXX
-    searches all droplets for the one with the given droplet_id
-    :param droplet_id: the one we're looking for
-    :return: that droplet, or None
-    """
-    def has_id(droplet):
-        return droplet.id == droplet_id
-
-    if droplet_id is None:
-        raise NonRecoverableError("droplet_id is required.")
-    else:
-        droplets = filter(has_id, ocean.Manager(token=load_token()).get_all_droplets())
-        sz = len(droplets)
-        if sz > 1:
-            msg = droplet_does_not_exist_for_operation("retrieve", droplet_id)
-            ctx.logger.debug(msg)
-            raise NonRecoverableError(msg)
-        elif sz == 1:
-            return droplets[0]
-        else:
-            return None
-
-
-def droplet_does_not_exist_for_operation(op, droplet_id):
-    """ Creates an error message when Droplets are unexpectedly not found for some operation
-    :param op: operation for which a Droplet does not exist
-    :param droplet_id: id that we expected to find
-    :return: a snotty message
-    """
-    return "Attempted to {0} a droplet with id '{1}', but no \
-    such Droplet exists in the system.".format(op, droplet_id)
-
-
 @operation
 def start(droplet_id=None, **_):
     """ XXX
@@ -161,7 +82,7 @@ def start(droplet_id=None, **_):
         start_droplet(create())
     else:
         ctx.logger.info("Starting existing droplet. Droplet id = '{0}'.".format(droplet_id))
-        d = get_droplet(droplet_id)
+        d = get_droplet(droplet_id, **_)
         if d is not None:
             start_droplet(d)
         else:
@@ -179,7 +100,7 @@ def stop(droplet_id, **_):
     :param droplet_id:
     :return: None
     """
-    d = get_droplet(droplet_id)
+    d = get_droplet(droplet_id, **_)
     if d is None:
         msg = droplet_does_not_exist_for_operation("stop", droplet_id)
         ctx.logger.debug(msg)
@@ -189,11 +110,3 @@ def stop(droplet_id, **_):
         d.destroy()
     # TODO need to check back later to see that the start operation has failed or succeeded or is still processing
     pass
-
-
-def main():
-    create()
-
-
-if __name__ == '__main__':
-    main()
